Remove old action entity
