--- packages/server/customer-os-api/entity/tenant.go
@@ -1,16 +1,0 @@
-package entity
-
-import (
-	neo4jentity "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/entity"
-	"time"
-)
-
-// Derepcated, use neo4jentity.TenantEntity instead
-type TenantEntity struct {
-	Id        string
-	Name      string
-	CreatedAt time.Time
-	UpdatedAt time.Time
-	Source    neo4jentity.DataSource
-	AppSource string
-}
