+++ packages/server/customer-os-api/service/action_service.go
@@ -2,20 +2,16 @@ package service
 
 import (
 	"context"
-	"github.com/neo4j/neo4j-go-driver/v5/neo4j/dbtype"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/entity"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/repository"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/common"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/logger"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/utils"
 	neo4jentity "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/entity"
 	neo4jenum "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/enum"
+	neo4jmapper "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/mapper"
 )
 
 type ActionService interface {
-	GetActionsForNodes(ctx context.Context, entityType neo4jenum.EntityType, ids []string) (*entity.ActionEntities, error)
-
-	mapDbNodeToActionEntity(dbNode dbtype.Node) *entity.ActionEntity
+	GetActionsForNodes(ctx context.Context, entityType neo4jenum.EntityType, ids []string) (*neo4jentity.ActionEntities, error)
 }
 
 type actionService struct {
@@ -30,32 +26,18 @@ func NewActionService(log logger.Logger, repository *repository.Repositories) Ac
 	}
 }
 
-func (s *actionService) GetActionsForNodes(ctx context.Context, entityType neo4jenum.EntityType, ids []string) (*entity.ActionEntities, error) {
+func (s *actionService) GetActionsForNodes(ctx context.Context, entityType neo4jenum.EntityType, ids []string) (*neo4jentity.ActionEntities, error) {
 	records, err := s.repositories.Neo4jRepositories.ActionReadRepository.GetFor(ctx, common.GetTenantFromContext(ctx), entityType, ids)
 	if err != nil {
 		return nil, err
 	}
 
-	var data entity.ActionEntities
+	var data neo4jentity.ActionEntities
 	for _, v := range records {
-		action := s.mapDbNodeToActionEntity(*v.Node)
+		action := neo4jmapper.MapDbNodeToActionEntity(v.Node)
 		action.DataloaderKey = v.LinkedNodeId
 		data = append(data, *action)
 	}
 
 	return &data, nil
-}
-
-func (s *actionService) mapDbNodeToActionEntity(dbNode dbtype.Node) *entity.ActionEntity {
-	props := utils.GetPropsFromNode(dbNode)
-	action := entity.ActionEntity{
-		Id:        utils.GetStringPropOrEmpty(props, "id"),
-		Type:      neo4jenum.GetActionType(utils.GetStringPropOrEmpty(props, "type")),
-		Content:   utils.GetStringPropOrEmpty(props, "content"),
-		Metadata:  utils.GetStringPropOrEmpty(props, "metadata"),
-		CreatedAt: utils.GetTimePropOrEpochStart(props, "createdAt"),
-		AppSource: utils.GetStringPropOrEmpty(props, "appSource"),
-		Source:    neo4jentity.GetDataSource(utils.GetStringPropOrEmpty(props, "source")),
-	}
-	return &action
 }
