+++ packages/server/customer-os-neo4j-repository/entity/action.go
@@ -7,6 +7,7 @@ import (
 )
 
 type ActionEntity struct {
+	DataLoaderKey
 	Id            string
 	CreatedAt     time.Time
 	Content       string
@@ -15,9 +16,9 @@ type ActionEntity struct {
 	Source        DataSource
 	SourceOfTruth DataSource
 	AppSource     string
-
-	DataloaderKey string
 }
+
+type ActionEntities []ActionEntity
 
 func (e *ActionEntity) GetDataloaderKey() string {
 	return e.DataloaderKey
@@ -32,4 +33,13 @@ func (ActionEntity) IsTimelineEvent() {
 
 func (ActionEntity) TimelineEventLabel() string {
 	return neo4jutil.NodeLabelAction
+}
+
+func (e *ActionEntity) Labels(tenant string) []string {
+	return []string{
+		neo4jutil.NodeLabelAction,
+		neo4jutil.NodeLabelAction + "_" + tenant,
+		neo4jutil.NodeLabelTimelineEvent,
+		neo4jutil.NodeLabelTimelineEvent + "_" + tenant,
+	}
 }
